+++ changelog.md
@@ -1,4 +1,4 @@
-﻿# Release Notes
+# Release Notes
 
 ## SmartStore.NET 3.1.5
 ### New Features
@@ -16,6 +16,7 @@
 * Moved StoreLastIpAddress & DisplayPrivacyAgreementOnContactUs from customer settings to privacy settings tab
 * #1450 Show the regular price only if it's higher than the final price
 * #1450 Do not ignore discounts with a negative amount
+* Set catalogsettings.showsharebutton to false as its not compliant with GDPR
 
 ### Bugfixes
 * Migration: take all same-named message templates into account